Add files via upload
In EDA, I am trying to plot a histogram for the number of occurences of images of each class. All the occurences of images for one data-batch is stored in the variable counter_dic array. Each of the entries of this array is a array itself labeled as [class_label_number: no. of occurences], where the class_label_number is the index of the class label we are referring to. Eg:- if class_label_number = 0, then label_names[0] = ‘frog’ etc. 

I saw a CNN implementation on internet for images, but would need more time to figure out the parameters of the implementation.  In the meantime, you guys can look at the code.